REACH-292 Ensure Python nodes build/run on Linux Reach
