+++ src/Dynamo.All.2012.sln
@@ -42,8 +42,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Arduino", "Libraries\Dynamo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Watch3D", "Libraries\DynamoWatch3D\Watch3D.csproj", "{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonUI", "Libraries\IronPython\IronPythonUI.csproj", "{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamoServices", "NodeServices\DynamoServices.csproj", "{EF879A10-041D-4C68-83E7-3192685F1BAE}"
 	ProjectSection(ProjectDependencies) = postProject
 		{133FC760-5699-46D9-BEA6-E816B5F01016} = {133FC760-5699-46D9-BEA6-E816B5F01016}
@@ -76,11 +74,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamoCoreWpfTests", "..\test\DynamoCoreWpfTests\DynamoCoreWpfTests.csproj", "{7DD8077A-201E-4C56-96C5-3C901A51BDF3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonTests", "..\test\Libraries\DynamoPythonTests\IronPythonTests.csproj", "{773988FE-EDF6-45CB-A63F-482955EB3553}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonEval", "Libraries\DSIronPython\IronPythonEval.csproj", "{9EEF4F42-6B3B-4358-9A8A-C2701539A822}"
-	ProjectSection(ProjectDependencies) = postProject
-		{133FC760-5699-46D9-BEA6-E816B5F01016} = {133FC760-5699-46D9-BEA6-E816B5F01016}
-	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FFITarget", "..\test\Engine\FFITarget\FFITarget.csproj", "{C70FE632-5500-4C57-B3D6-9B5574137551}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -251,10 +244,6 @@ Global
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -287,10 +276,6 @@ Global
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -458,8 +443,6 @@ Global
 		{F5932F7D-8E34-4787-80B8-E7F9D996EDFF} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{9589D7A5-12BE-446B-BBFF-B8C4951EDB13} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{398542E6-659A-48C8-86EB-33164D227C90} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{6E0A079E-85F1-45A1-AD5B-9855E4344809} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{9B4FDC96-E2F9-4B8F-894A-4294405D50E7} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
