+++ src/Dynamo.All.2013.sln
@@ -50,8 +50,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Arduino", "Libraries\Dynamo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Watch3D", "Libraries\DynamoWatch3D\Watch3D.csproj", "{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonUI", "Libraries\IronPython\IronPythonUI.csproj", "{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamoServices", "NodeServices\DynamoServices.csproj", "{EF879A10-041D-4C68-83E7-3192685F1BAE}"
 	ProjectSection(ProjectDependencies) = postProject
 		{133FC760-5699-46D9-BEA6-E816B5F01016} = {133FC760-5699-46D9-BEA6-E816B5F01016}
@@ -84,11 +82,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamoCoreWpfTests", "..\test\DynamoCoreWpfTests\DynamoCoreWpfTests.csproj", "{7DD8077A-201E-4C56-96C5-3C901A51BDF3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonTests", "..\test\Libraries\DynamoPythonTests\IronPythonTests.csproj", "{773988FE-EDF6-45CB-A63F-482955EB3553}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IronPythonEval", "Libraries\DSIronPython\IronPythonEval.csproj", "{9EEF4F42-6B3B-4358-9A8A-C2701539A822}"
-	ProjectSection(ProjectDependencies) = postProject
-		{133FC760-5699-46D9-BEA6-E816B5F01016} = {133FC760-5699-46D9-BEA6-E816B5F01016}
-	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FFITarget", "..\test\Engine\FFITarget\FFITarget.csproj", "{C70FE632-5500-4C57-B3D6-9B5574137551}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -254,10 +247,6 @@ Global
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EF879A10-041D-4C68-83E7-3192685F1BAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -290,10 +279,6 @@ Global
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{773988FE-EDF6-45CB-A63F-482955EB3553}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C70FE632-5500-4C57-B3D6-9B5574137551}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -460,7 +445,6 @@ Global
 		{F5932F7D-8E34-4787-80B8-E7F9D996EDFF} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{9589D7A5-12BE-446B-BBFF-B8C4951EDB13} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{B9E338CA-6677-4772-B01D-1FCEABCDAAAB} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
-		{01DE9B06-0BCB-4D8A-862E-E8170F5D6B4F} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{0355EB79-6961-4B8A-8F66-035D4EC0C9FE} = {F4D44BC0-32CF-4E58-AD2A-F19CE1450B00}
 		{6E1177DB-5E4F-4E2E-82E7-902437E56AED} = {0355EB79-6961-4B8A-8F66-035D4EC0C9FE}
 		{2235F7CA-AB25-4177-92BC-4B574D2D5DF6} = {0355EB79-6961-4B8A-8F66-035D4EC0C9FE}
@@ -470,7 +454,6 @@ Global
 		{472084ED-1067-4B2C-8737-3839A6143EB2} = {24A1EF6B-5A24-4536-B8BD-C879ABA26389}
 		{7DD8077A-201E-4C56-96C5-3C901A51BDF3} = {24A1EF6B-5A24-4536-B8BD-C879ABA26389}
 		{773988FE-EDF6-45CB-A63F-482955EB3553} = {0E492D35-2310-4849-9694-A2A53C09F21B}
-		{9EEF4F42-6B3B-4358-9A8A-C2701539A822} = {FA7BE306-A3B0-45FA-9D87-0C69E6932C13}
 		{C70FE632-5500-4C57-B3D6-9B5574137551} = {0355EB79-6961-4B8A-8F66-035D4EC0C9FE}
 		{88A44558-9148-4F1C-876B-46472793F3D6} = {F4D44BC0-32CF-4E58-AD2A-F19CE1450B00}
 		{1254D922-643B-4F0B-82BD-E9E02399B040} = {88A44558-9148-4F1C-876B-46472793F3D6}
