+++ src/Libraries/CoreNodeModels/Properties/Resources.resx
@@ -533,4 +533,35 @@
   <data name="SelectionNodeSugestion" xml:space="preserve">
     <value>Select something in the model.</value>
   </data>
+  <data name="PythonNodePortDataOutputToolTip" xml:space="preserve">
+    <value>Result of the python script</value>
+  </data>
+  <data name="PythonScriptDescription" xml:space="preserve">
+    <value>Runs an embedded IronPython script.</value>
+    <comment>Description for Python Script</comment>
+  </data>
+  <data name="PythonScriptEditoCancelButton" xml:space="preserve">
+    <value>Cancel</value>
+  </data>
+  <data name="PythonScriptEditorAcceptChangesButton" xml:space="preserve">
+    <value>Accept Changes</value>
+  </data>
+  <data name="PythonScriptEditorCaption" xml:space="preserve">
+    <value>Edit Python Script...</value>
+  </data>
+  <data name="PythonScriptEditorInputComment" xml:space="preserve">
+    <value>The inputs to this node will be stored as a list in the IN variables.</value>
+    <comment>It is comment in the code. IN shouldn't be translated.</comment>
+  </data>
+  <data name="PythonScriptEditorOutputComment" xml:space="preserve">
+    <value>Assign your output to the OUT variable.</value>
+    <comment>It is comment in the code. OUT shouldn't be translated.</comment>
+  </data>
+  <data name="PythonScriptFromStringDescription" xml:space="preserve">
+    <value>Runs a IronPython script from a string.</value>
+    <comment>Description for Python Script From String</comment>
+  </data>
+  <data name="PythonStringPortDataScriptToolTip" xml:space="preserve">
+    <value>Python script to run.</value>
+  </data>
 </root>
