+++ src/Libraries/CoreNodeModels/PythonNode.cs
@@ -2,66 +2,20 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Controls;
-using System.Windows.Input;
+
 using System.Xml;
+
+using Autodesk.DesignScript.Runtime;
+
 using DSIronPython;
-using Dynamo.Controls;
-using Dynamo.Core;
+
 using Dynamo.Models;
 using Dynamo.Nodes;
-using Dynamo.Utilities;
-using Dynamo.ViewModels;
-using Dynamo.Wpf;
 
 using ProtoCore.AST.AssociativeAST;
-using Autodesk.DesignScript.Runtime;
-using ProtoCore.Namespace;
 
-namespace DSIronPythonNode
+namespace DSCoreNodesUI
 {
-    public class PythonNodeViewCustomization : VariableInputNodeViewCustomization, INodeViewCustomization<PythonNode>
-    {
-        private DynamoViewModel dynamoViewModel;
-        private PythonNode model;
-
-        public void CustomizeView(PythonNode nodeModel, NodeView nodeView)
-        {
-            base.CustomizeView(nodeModel, nodeView);
-
-            model = nodeModel;
-            dynamoViewModel = nodeView.ViewModel.DynamoViewModel;
-
-            var editWindowItem = new MenuItem { Header = "Edit...", IsCheckable = false };
-            nodeView.MainContextMenu.Items.Add(editWindowItem);
-            editWindowItem.Click += delegate { EditScriptContent(); };
-            nodeView.UpdateLayout();
-
-            nodeView.MouseDown += view_MouseDown;
-        }
-
-        private void view_MouseDown(object sender, MouseButtonEventArgs e)
-        {
-            if (e.ClickCount >= 2)
-            {
-                EditScriptContent();
-                e.Handled = true;
-            }
-        }
-
-        private void EditScriptContent()
-        {
-            var editWindow = new ScriptEditorWindow(dynamoViewModel);
-            editWindow.Initialize(model.GUID, "ScriptContent", model.Script);
-            bool? acceptChanged = editWindow.ShowDialog();
-            if (acceptChanged.HasValue && acceptChanged.Value)
-            {
-                // Mark node for update
-                model.OnNodeModified();
-            }
-        }
-    }
-
     public abstract class PythonNodeBase : VariableInputNode
     {
         protected PythonNodeBase()
@@ -110,7 +64,7 @@ namespace DSIronPythonNode
 
     [NodeName("Python Script")]
     [NodeCategory(BuiltinNodeCategories.CORE_SCRIPTING)]
-    [NodeDescription("PythonScriptDescription", typeof(DSIronPythonNode.Properties.Resources))]
+    [NodeDescription("PythonScriptDescription", typeof(Properties.Resources))]
     [SupressImportIntoVM]
     [IsDesignScriptCompatible]
     public sealed class PythonNode : PythonNodeBase
@@ -199,7 +153,7 @@ namespace DSIronPythonNode
 
     [NodeName("Python Script From String")]
     [NodeCategory(BuiltinNodeCategories.CORE_SCRIPTING)]
-    [NodeDescription("PythonScriptFromStringDescription", typeof(DSIronPythonNode.Properties.Resources))]
+    [NodeDescription("PythonScriptFromStringDescription", typeof(Properties.Resources))]
     [SupressImportIntoVM]
     [IsDesignScriptCompatible]
     public sealed class PythonStringNode : PythonNodeBase
@@ -234,5 +188,4 @@ namespace DSIronPythonNode
             };
         }
     }
-
 }
