+++ src/Libraries/CoreNodeModelsWpf/CoreNodeModelsWpf.csproj
@@ -38,8 +38,25 @@
       <HintPath>..\..\..\extern\Helix3D\NET40\HelixToolkit.Wpf.SharpDX.dll</HintPath>
       <Private>False</Private>
     </Reference>
+    <Reference Include="ICSharpCode.AvalonEdit">
+      <HintPath>..\..\..\extern\avalonEdit\ICSharpCode.AvalonEdit.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="IronPython">
+      <HintPath>..\..\..\extern\IronPython\IronPython.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Microsoft.Dynamic">
+      <HintPath>..\..\..\extern\IronPython\Microsoft.Dynamic.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
     <Reference Include="Microsoft.Practices.Prism">
       <HintPath>..\..\..\extern\prism\Microsoft.Practices.Prism.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Microsoft.Scripting">
+      <HintPath>..\..\..\extern\IronPython\Microsoft.Scripting.dll</HintPath>
+      <Private>False</Private>
     </Reference>
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
@@ -70,8 +87,11 @@
       <DependentUpon>DynamoSlider.xaml</DependentUpon>
     </Compile>
     <Compile Include="ConverterViewModel.cs" />
+    <Compile Include="IronPythonCompletionData.cs" />
+    <Compile Include="IronPythonCompletionProvider.cs" />
     <Compile Include="NodeViewCustomizations\ConverterNodeViewCustomization.cs" />
     <Compile Include="NodeViewCustomizations\DoubleInput.cs" />
+    <Compile Include="NodeViewCustomizations\PythonNode.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="NodeViewCustomizations\BoolSelector.cs" />
     <Compile Include="NodeViewCustomizations\ColorRange.cs" />
@@ -99,6 +119,9 @@
     <Compile Include="NodeViewCustomizations\StringInput.cs" />
     <Compile Include="NodeViewCustomizations\WatchImage.cs" />
     <Compile Include="NodeViewCustomizations\Watch.cs" />
+    <Compile Include="Controls\ScriptEditorWindow.xaml.cs">
+      <DependentUpon>ScriptEditorWindow.xaml</DependentUpon>
+    </Compile>
     <Compile Include="SliderViewModel.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -154,12 +177,27 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
+    <Page Include="Controls\ScriptEditorWindow.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\DeprecatedNode.png" />
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\MissingNode.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\class.png" />
+    <EmbeddedResource Include="Resources\field.png" />
+    <EmbeddedResource Include="Resources\method.png" />
+    <EmbeddedResource Include="Resources\namespace.png" />
+    <EmbeddedResource Include="Resources\privateMethod.png" />
+    <EmbeddedResource Include="Resources\property.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\ICSharpCode.PythonBinding.Resources.Python.xshd" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
