+++ src/Libraries/CoreNodeModelsWpf/IronPythonCompletionProvider.cs
