+++ src/Libraries/CoreNodeModelsWpf/Resources/ICSharpCode.PythonBinding.Resources.Python.xshd
