+++ src/Libraries/CoreNodeModelsWpf/Resources/method.png
