+++ src/Libraries/CoreNodeModelsWpf/Resources/namespace.png
