+++ src/Libraries/CoreNodeModelsWpf/Resources/privateMethod.png
