--- src/Libraries/DSIronPython/DSIronPython_DynamoCustomization.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>DSIronPython</name>
-    </assembly>
-    <namespaces>
-        <namespace name="DSIronPython">
-            <category>Core.Scripting</category>
-        </namespace>
-    </namespaces>
-</doc>
