--- src/Libraries/DSIronPython/Properties/AssemblyInfo.cs
@@ -1,11 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("DSIronPython")]
-[assembly: AssemblyCulture("")]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("2cbb6e0f-dfca-49dc-8cff-bc65d82951a1")]
