--- src/Libraries/IronPython/Properties/AssemblyInfo.cs
@@ -1,11 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-
-[assembly: AssemblyTitle("DSIronPythonNode")]
-[assembly: AssemblyCulture("")]
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("5df40db3-627c-43fb-9a0e-0ee422d92485")]
