+++ test/DynamoCoreTests/DSEvaluationModelTest.cs
@@ -18,7 +18,6 @@ namespace Dynamo.Tests
         {
             libraries.Add("ProtoGeometry.dll");
             libraries.Add("DSCoreNodes.dll");
-            libraries.Add("DSIronPython.dll");
             libraries.Add("FunctionObject.ds");
             base.GetLibrariesToPreload(libraries);
         }
