+++ test/DynamoCoreWpfTests/DynamoCoreWpfTests.csproj
@@ -169,11 +169,6 @@
       <Name>DynamoConversions</Name>
       <Private>False</Private>
     </ProjectReference>
-    <ProjectReference Include="..\..\src\Libraries\IronPython\IronPythonUI.csproj">
-      <Project>{01de9b06-0bcb-4d8a-862e-e8170f5d6b4f}</Project>
-      <Name>IronPythonUI</Name>
-      <Private>False</Private>
-    </ProjectReference>
     <ProjectReference Include="..\..\src\Libraries\UnitsUI\UnitsUI.csproj">
       <Project>{c67a7e80-73e6-4bbf-9d3f-dcd86ce306be}</Project>
       <Name>UnitsUI</Name>
