+++ test/Libraries/DynamoPythonTests/PythonEditTests.cs
@@ -1,7 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using DSIronPythonNode;
+
+using DSCoreNodesUI;
+
 using Dynamo.Models;
 using NUnit.Framework;
 using DynCmd = Dynamo.Models.DynamoModel;
@@ -12,7 +14,7 @@ namespace Dynamo.Tests
     {
         protected override void GetLibrariesToPreload(List<string> libraries)
         {
-            libraries.Add("DSIronPython.dll");
+            libraries.Add("DSCoreNodes.dll");
             base.GetLibrariesToPreload(libraries);
         }
 
@@ -27,7 +29,7 @@ namespace Dynamo.Tests
             // get the python node
             var workspace = model.CurrentWorkspace;
             var nodeModel = workspace.NodeFromWorkspace("3bcad14e-d086-4278-9e08-ed2759ef92f3");
-            DSIronPythonNode.PythonNode pynode = nodeModel as DSIronPythonNode.PythonNode;
+            var pynode = nodeModel as PythonNode;
             Assert.NotNull(pynode);
 
             // make changes to python script
