+++ test/Libraries/WorkflowTests/PackageValidationTest.cs
@@ -17,7 +17,6 @@ namespace Dynamo.Tests
         {
             libraries.Add("ProtoGeometry.dll");
             libraries.Add("DSCoreNodes.dll");
-            libraries.Add("DSIronPython.dll");
             libraries.Add("FunctionObject.ds");
             base.GetLibrariesToPreload(libraries);
         }
